buzzaar-backend/buzzaar_core/pyproject.toml
[tool.poetry]
name = "buzzaar_core"
version = "0.1.0"
description = "A comprehensive C2C marketplace platform with match-to-buy feature."
authors = ["codelotusr <lunarcode0@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.8.1"
attrs = "24.2.0"
certifi = "2024.8.30"
cffi = "1.17.1"
channels = "4.1.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "43.0.1"
cssbeautifier = "1.15.1"
dj-rest-auth = "6.0.0"
django = "5.1.1"
django-allauth = "65.0.2"
django-axes = "7.0.0"
django-cors-headers = "4.4.0"
django-environ = "0.11.2"
django-filter = "24.3"
django-hijack = "3.6.1"
django-jsonfield = "1.4.1"
django-oauth-toolkit = "3.0.1"
django-rest-framework = "0.1.0"
django-storages = "1.14.4"
django-stripe-payments = "2.0.0"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
editorconfig = "0.12.4"
factory-boy = "3.3.1"
faker = "30.3.0"
html-tag-names = "0.1.2"
html-void-elements = "0.1.0"
idna = "3.10"
inflection = "0.5.1"
iniconfig = "2.0.0"
isort = "5.13.2"
jsbeautifier = "1.15.1"
json5 = "0.9.25"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jwcrypto = "1.5.6"
oauthlib = "3.2.2"
packaging = "24.1"
pathspec = "0.12.1"
pillow = "11.0.0"
pluggy = "1.5.0"
pycparser = "2.22"
pydot = "3.0.2"
pyjwt = "2.9.0"
pyparsing = "3.2.0"
pytest = "8.3.3"
pytest-django = "4.9.0"
pytest-factoryboy = "2.7.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.2"
pyyaml = "6.0.2"
referencing = "0.35.1"
regex = "2024.9.11"
requests = "2.32.3"
rpds-py = "0.20.1"
six = "1.16.0"
sqlparse = "0.5.1"
stripe = "11.1.0"
tqdm = "4.66.5"
types-psycopg2 = "2.9.21.20241019"
types-pyyaml = "6.0.12.20240917"
typing-extensions = "4.12.2"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
psycopg2-binary = "^2.9.10"
pytest-cov = "^6.0.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "4.4.6"
django-stubs = "5.1.1"
django-stubs-ext = "5.1.1"
django-extensions = "3.2.3"
djhtml = "3.0.6"
djlint = "1.35.2"
drf-spectacular = "0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
